requirements.txt
azure-servicebus
openai
asyncio
